% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummyvars}
\alias{dummyvars}
\title{Create dummy variables for all categorical variables in a data.frame. Can be
used as a pre-processing step before \code{\link{pm_input_info}}.}
\usage{
dummyvars(df)
}
\arguments{
\item{df}{a data.frame on which to make dummy variables for each
categorical/factor variable, based on contrasts.}
}
\value{
a data.frame in which the new columns are appended to the \code{df}.
Naming convention of the new dummy variables is variable_level. For
example, a factor variable called "colour" with levels "red", "green" and
"purple" (where "purple" is the reference) will have two new "dummy
variables" named colour_red and colour_green.
}
\description{
Create dummy variables for all categorical variables in a data.frame. Can be
used as a pre-processing step before \code{\link{pm_input_info}}.
}
\examples{
dummyvars(data.frame("Colour" = factor(sample(c("red",
                                                "azure",
                                                "green",
                                                "white"),
                                             500,
                                             replace = TRUE))))
}
