% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_calculate.R
\name{pm_calculate}
\alias{pm_calculate}
\title{Apply an existing prediction model to new data, to obtain predicted risks,
given published coefficients}
\usage{
pm_calculate(
  formula,
  type = c("logistic", "survival"),
  existingcoefs,
  newdata,
  add = FALSE
)
}
\arguments{
\item{formula}{an object of class "\code{\link[stats]{formula}}" (or a
character string that can be coerced to that class). This is a symbolic
description of the existing prediction model. The details of the model
specification are given under "Details".}

\item{type}{specifies the type of model that the existing prediction model is
based on; possible options are:
\itemize{
\item {\code{"logistic"} indicates that the existing model was based
on a logistic regression model}
\item {\code{"survival"} indicates that the existing model was based
on a survival regression model}
}
\strong{Currently, only \code{"logistic"} is supported.}}

\item{existingcoefs}{a named vector of coefficients, taken exactly as
published from the existing prediction model. Names much match those in
\code{newdata}. See "Details".}

\item{newdata}{a data frame in which to look for variables with which to
apply the existing prediction model to (i.e. predict risk). Variable names
must match those in \code{existingcoefs}. \strong{Currently assumes all
data pre-processing has been done before calling function}.}

\item{add}{Logical variable indicating if the calculated predictions
should be added as new columns to newdata (TRUE) or should be returned
separately (FALSE, default).}
}
\value{
Linear predictor and predicted risks for each observation in
\code{newdata}, based on the specified information about the existing
prediction model
}
\description{
Apply an existing prediction model to new data, to obtain predicted risks,
given published coefficients
}
\details{
The \code{formula} argument must be specified in the form
\code{outcome ~ predictors}, where \code{outcome} is the name of the
column in \code{newdata} that stores the binary outcome values (0 or 1),
and \code{predictors} is the functional form of the existing prediction
model, where each predictor is separated with a \code{+} symbol.
}
