% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_predict.R
\name{pm_predict}
\alias{pm_predict}
\title{Use an existing prediction model to estimate predicted risks in new data,
given published coefficients}
\usage{
pm_predict(x, time_horizon = NULL)
}
\arguments{
\item{x}{an object of class "\code{pminfo}" produced by
\code{\link{pm_input_info}}. This is a 'blueprint' description of the
existing prediction model, and the newdata on which predictions should be
made.}

\item{time_horizon}{for survival models, an integer giving the time horizon at
which a prediction is required.
Currently, must match a time in x$baselinehazard.}
}
\value{
Linear predictor and predicted risks for each observation in
\code{newdata}, based on the specified information about the existing
prediction model
}
\description{
Use an existing prediction model to estimate predicted risks in new data,
given published coefficients
}
\details{
This function takes the relevant information about the existing
prediction model (as supplied by called \code{\link{pm_input_info}}), and
returns the predicted risks for each individual/observation in
\code{newdata}. See \code{\link{pm_input_info}}) for more details.
}
\seealso{
\code{\link{pm_input_info}}
}
