% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_validate_survival.R
\name{pm_validate_survival}
\alias{pm_validate_survival}
\alias{pm_validate_survival.data.frame}
\alias{pm_validate_survival.pminfo}
\title{Validate an existing prediction model based on survival/time-to-event models,
to calculate the predictive performance of an existing prediction model,
against a new dataset.}
\usage{
pm_validate_survival(x, ...)

\method{pm_validate_survival}{data.frame}(x, LinearPredictor, Outcome, ...)

\method{pm_validate_survival}{pminfo}(x, ...)
}
\arguments{
\item{x}{an object of class "pminfo", or a data.frame. If a data.frame is
supplied, \code{LinearPredictor} and \code{Outcome} should also be supplied
as described below.}

\item{...}{other arguments}

\item{LinearPredictor}{character variable specifying the column name that
stores the linear predictor of the model to be validated}

\item{Outcome}{character variable specifying the column name that
stores the observed outcomes within the validation dataset}
}
\value{
TO ADD
}
\description{
Validate an existing prediction model based on survival/time-to-event models,
to calculate the predictive performance of an existing prediction model,
against a new dataset.
}
\examples{
#TO ADD

}
\seealso{
\code{\link{pm_input_info}}
}
