% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_stacked_regression.R
\name{pred_stacked_regression}
\alias{pred_stacked_regression}
\title{Perform Stacked Regression on Existing Prediction Models}
\usage{
pred_stacked_regression(x, positivity_constraint = FALSE, ...)
}
\arguments{
\item{x}{an object of class "\code{predinfo}" produced by calling
\code{\link{pred_input_info}} containing information on at least two
existing prediction models.}

\item{positivity_constraint}{TRUE/FALSE denoting if the weights within the
stacked regression model should be constrained to be non-negative (TRUE) or
should be allowed to take any value (FALSE). See details}
}
\value{
A object of class "\code{predSR}". This is a list containing the
information/estimates about the meta-model obtained by stacked regression.
Specifically, \itemize{ \item{Stacked_Regression_Model = the results of the
stacked regression equation} \item{Aggregated_Coefs = a data.frame of the
pooled/aggregated coefficients of each covariate across the existing
prediction models} \item{pred_info = the original "\code{predinfo}" object,
as passed to \code{x}}}
}
\description{
This function takes a set of existing prediction models, and uses the new
dataset to combine/aggregate them into a single 'meta-model', as described in
Debray et al. 2014.
}
\details{
The aim of this function is to take a set of (previously estimated)
prediction models that were each originally developed for the same
prediction task, and pool/aggregate these into a single prediction model
(meta-model) using stacked regression based on new data (data not used to
develop any of the existing models). The methodological details can be
found in Debray et al. 2014.

Given that the existing models are likely to be highly co-linear (given
that they were each developed for the same prediction task), it has been
proposed to impose a positivity constraint on the weights of the stacked
regression model (Debray et al. 2014.). This is not strictly needed, so the
use of this constraint can be controlled using
\code{positivity_constraint}. If \code{positivity_constraint} is set to
TRUE, then the stacked regression model will be estimated by optimising the
(log-)likelihood of a logistic regression model using bound constrained
optimization ("L-BFGS-B" in \code{stats::optim()}).
}
\examples{
model_input <- pred_input_info(model_type = "logistic",
                               model_info = SYNPM$Existing_models,
                               newdata = SYNPM$ValidationData,
                               binary_outcome = "Y")
SR <- pred_stacked_regression(model_input)
summary(SR)
#one could then validate this as follows (but note this should be adjusted for in-sample optimism)
pred_validate(pred_input_info(model_type = "logistic",
                              model_info = SR$Aggregated_Coefs,
                              newdata = SYNPM$ValidationData,
                              binary_outcome = "Y"))

}
\references{
Debray, T.P., Koffijberg, H., Nieboer, D., Vergouwe, Y.,
Steyerberg, E.W. and Moons, K.G. (2014), Meta-analysis and aggregation of
multiple published prediction models. \emph{Statistics in Medicine}, 33:
2341-2362
}
