% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRupdate-package.R
\docType{package}
\name{predRupdate-package}
\alias{predRupdate}
\alias{predRupdate-package}
\title{predRupdate: Prediction Model Validation and Updating}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Evaluate the predictive performance of an existing (i.e. previously developed) prediction/ prognostic model given relevant information about the existing prediction model (e.g. coefficients) and a new dataset. Provides a range of model updating methods that help tailor the existing model to the new dataset; see Su et al. (2018) \doi{10.1177/0962280215626466}. Techniques to aggregate multiple existing prediction models on the new data are also provided; see Debray et al. (2014) \doi{10.1002/sim.6080} and Martin et al. (2018) \doi{10.1002/sim.7586}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GlenMartin31/predRupdate}
  \item \url{https://glenmartin31.github.io/predRupdate/}
  \item Report bugs at \url{https://github.com/GlenMartin31/predRupdate/issues}
}

}
\author{
\strong{Maintainer}: Glen P. Martin \email{glen.martin31@gmail.com} (\href{https://orcid.org/0000-0002-3410-9472}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item David Jenkins \email{david.jenkins-5@manchester.ac.uk} [copyright holder]
  \item Matthew Sperrin \email{matthew.sperrin@manchester.ac.uk} [copyright holder]
}

}
\keyword{internal}
