% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_probabilities.R
\name{validate_probabilities}
\alias{validate_probabilities}
\title{Validate predicted probabilities against binary outcome}
\usage{
validate_probabilities(ObservedOutcome, Prob, Logit)
}
\arguments{
\item{ObservedOutcome}{a vector of N binary observations, denoting if the
outcome was observed (1) or not observed (0) for each individual in the
validation dataset}

\item{Prob}{a vector of N predicted probabilities from the model. Specify
either \code{Prob} or \code{Logit}.}

\item{Logit}{a vector of N observations where each is the calculated linear
predictor (log-odds) from the existing model that is being evaluated.
Specify either \code{Prob} or \code{Logit}.}
}
\value{
Returns a list of the performance metrics and associated 95\%
confidence intervals, where appropriate.
}
\description{
This function is used to validate predicted probabilities (usually from an
existing/ previously developed logistic regression model) against binary
outcomes. It takes a vector of predicted risks (or the log odds - the linear
predictor) and a vector of binary observed probabilities (not used to develop
the model that generates the predictions). From which, the function
calculates metrics of calibration, discrimination and overall accuracy.
}
\details{
TO ADD
}
\seealso{
\code{\link{pm_validate}}
}
