% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummyvars}
\alias{dummyvars}
\title{Create dummy variables for all categorical variables in a data.frame. Can be
used as a pre-processing function within \code{\link{pm_input_info}}.}
\usage{
dummyvars(df)
}
\arguments{
\item{df}{a data.frame on which to make dummy variables for each
categorical/factor variable, based on contrasts.}
}
\value{
a list in which each element is a dummy variable for each categorical
variable in \code{df}. All combinations of dummy variable are returned.
Naming convention of the new dummy variables is variable_level. For
example, a factor variable in \code{df} named "colour" with levels "red",
"green" and "purple" (where "purple" is the reference) will produce a list
with two elements (the new dummy variables), named colour_red and
colour_green, respectively.
}
\description{
Create dummy variables for all categorical variables in a data.frame. Can be
used as a pre-processing function within \code{\link{pm_input_info}}.
}
\examples{
dummyvars(data.frame("Colour" = factor(sample(c("red",
                                                "azure",
                                                "green",
                                                "white"),
                                             500,
                                             replace = TRUE))))
}
\seealso{
\code{\link{pm_input_info}}
}
\author{
Glen Martin, \email{glen.martin@manchester.ac.uk}
}
