% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_validate.R
\name{pm_validate}
\alias{pm_validate}
\alias{pm_validate.data.frame}
\title{Validate an existing prediction}
\usage{
pm_validate(x, ...)

\method{pm_validate}{data.frame}(
  x,
  model_type = c("logistic", "survival"),
  LinearPredictor = NULL,
  binary_outcome = NULL,
  survival_time = NULL,
  event_indicator = NULL,
  ...
)
}
\arguments{
\item{x}{an object used to select a method - should be of class "pminfo" or
"data.frame". If a data.frame this should include the linear predictor of
the existing prediction model and the observed outcomes to validate this
model against.}

\item{...}{further arguments passed to other methods}

\item{model_type}{specifies the type of model that the existing prediction
model is based on; possible options are: \itemize{ \item {\code{"logistic"}
indicates that the existing model was based on a logistic regression model
(default)} \item {\code{"survival"} indicates that the existing model was
based on a survival regression model} }}

\item{LinearPredictor}{character variable specifying the column of \code{x}
that stores the linear predictor of the model to be validated}

\item{binary_outcome}{Character variable giving the name of the column in
\code{x} that represents the observed binary outcomes. Only relevant for
\code{model_type}="logistic"; leave as default \code{NULL} otherwise.}

\item{survival_time}{Character variable giving the name of the column in
\code{x} that represents the observed survival times. Only relevant for
\code{model_type}="survival"; leave as default \code{NULL} otherwise.}

\item{event_indicator}{Character variable giving the name of the column in
\code{x} that represents the observed survival indicator (1 for event, 0
for censoring). Only relevant for \code{model_type}="survival"; leave as
default \code{NULL} otherwise.}
}
\value{
TO ADD
}
\description{
Validate an existing prediction model, to calculate the predictive
performance against a new (validation) dataset.
}
\examples{
#TO ADD

}
\seealso{
\code{\link{pm_input_info}}
}
