% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_validate.R
\name{pm_validate}
\alias{pm_validate}
\title{Validate an existing prediction}
\usage{
pm_validate(x, ...)
}
\arguments{
\item{x}{an object of class "pminfo"}

\item{...}{further arguments passed to other methods}
}
\value{
TO ADD
}
\description{
Validate an existing prediction model, to calculate the predictive
performance against a new (validation) dataset.
}
\examples{
#TO ADD

}
\seealso{
\code{\link{pm_input_info}}
}
