% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_predict.R
\name{pred_predict}
\alias{pred_predict}
\title{Make predictions from an existing prediction model}
\usage{
pred_predict(x, time_horizon = NULL)
}
\arguments{
\item{x}{an object of class "\code{predinfo}" produced by calling
\code{\link{pred_input_info}}.}

\item{time_horizon}{for survival models, an integer giving the time horizon
(post baseline/time of prediction) at which a prediction is required.
Currently, this must match a time in x$baselinehazard.}
}
\value{
\code{\link{pred_predict}} returns a list containing the following
components: \itemize{ \item{LinearPredictor = the linear predictor for each
observation in the new data (i.e., the linear combination of the models
predictor variables and their corresponding coefficients)}
\item{PredictedRisk = the predicted risk for each observation in the new
data} \item{TimeHorizon = for survival models, an integer giving the time
horizon at which a prediction is made} \item{Outcomes = vector of
outcomes/endpoints (if avaliable).} }
}
\description{
Use an existing prediction model to estimate predicted risks of the outcome
for each observation in a new dataset.
}
\details{
This function takes the relevant information about the existing
prediction model (as supplied by calling \code{\link{pred_input_info}}),
and returns the predicted risks for each individual/observation in
\code{newdata}. See \code{\link{pred_input_info}}) for more details.

If the existing prediction model is based on logistic regression (i.e., if
x$model_type == "logistic"), this will be the predicted probability of the
binary outcome conditional on the predictor variables in the newdatas
(i.e., \eqn{P(Y=1 | X)}). If the existing prediction model is based on a
time-to-event/survival model (i.e., if x$model_type == "survival"), this
will be one minus the survival probability (i.e., \eqn{1 - S(T>time_horizon
  | X)}).
}
\seealso{
\code{\link{pred_input_info}}
}
