% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_validate.R
\name{pred_validate}
\alias{pred_validate}
\title{Validate an existing prediction}
\usage{
pred_validate(x, ...)
}
\arguments{
\item{x}{an object of class "predinfo"}

\item{...}{further arguments passed to other methods. See
\code{\link{validate_probabilities}} for more details of arguments that can
be passed}
}
\value{
A list of performance metrics, estimated by applying the existing
prediction model to the newdata.
}
\description{
Validate an existing prediction model, to calculate the predictive
performance against a new (validation) dataset.
}
\details{
This function takes an existing prediction model formatted according
to \code{\link{pred_input_info}}, and calculates measures of predictive
performance on new data (e.g., within an external validation study). The
information about the existing prediction model should first be inputted by
calling \code{\link{pred_input_info}}, before passing the resulting object
to \code{pred_validate}.

In the case of validating a logistic regression model, \code{pred_validate}
internally calls \code{\link{validate_probabilities}}. See this function
for details on the performance metrics that are estimated.

In the case of validating a survival prediction model,...
}
\examples{
# Example 1 - logistic regression example
existing_cpm_info <- pred_input_info(model_type = "logistic",
                                     existingcoefs = c("(Intercept)" = -3.0893961710923,
                                                       "Age" = 0.0230955938292795,
                                                       "SexM" = 0.263578567485447,
                                                       "Smoking_Status" = 0.689825139075564,
                                                       "Diabetes" = 0.387810349702088,
                                                       "CKD" = 0.56129156010678),
                                     formula = formula(SYNPM$Existing_models$Formula[2]),
                                     newdata = SYNPM$ValidationData,
                                     pre_processing = list(function(df) {dummyvars(df)}),
                                     binary_outcome = "Y")
pred_validate(existing_cpm_info)

# Example 2 - survival example
#TO ADD

}
\seealso{
\code{\link{pred_input_info}} \code{\link{validate_probabilities}}
}
