% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMART}
\alias{SMART}
\title{Second Manifestations of ARTerial disease (SMART) dataset}
\format{
A data with 3873 rows (observations) with 16 variables:
\describe{
\item{TEVENT}{Time of event (days)}
\item{EVENT}{Cardiovascular event (0=no, 1=yes)}
\item{SEX}{M = male, f = female}
\item{AGE}{Age (years)}
\item{DIABETES}{History of diabetes (0=no, 1=yes)}
\item{CEREBRAL}{History of cerebrovascular disease (0=no, 1=yes)}
\item{CARDIAC}{History of cardiovascular disease (0=no, 1=yes)}
\item{AAA}{History of abdominal aortic aneurysm (0=no, 1=yes)}
\item{PERIPH}{History of periferal aortic aneurysm (0=no, 1=yes)}
\item{STENOSIS}{History of periferal vascular disease (0=no, 1=yes)}
\item{SYSTBP}{Systolic blood pressure (mm Hg)}
\item{DIASTBP}{Diastolic blood pressure (mm Hg)}
\item{BMIO}{Body mass index (kg/m^2)}
\item{CHOLO}{Cholesterol level (mmol/L)}
\item{albumin}{Albumin in urine (no/low/high)}
\item{SMOKING}{Smoking status (never/former/current)}
}
}
\source{
\url{http://www.clinicalpredictionmodels.org/doku.php?id=rcode_and_data:start;
https://doi.org/10.1023/A:1007621514757}
}
\usage{
SMART
}
\description{
The Second Manifestations of ARTerial disease (SMART) study is a
single-centre prospective cohort study among patients, newly referred to the
hospital with (1) clinically manifest atherosclerotic vessel disease, or (2)
marked risk factors for atherosclerosis. Further details on the data can be
found here: Simons, P., Algra, A., van de Laak, M. et al. Second
Manifestations of ARTerial disease (SMART) study: Rationale and design. Eur J
Epidemiol 15, 773–781 (1999). https://doi.org/10.1023/A:1007621514757.
}
\details{
Interest is in predicting the occurrence of cardiovascular events; this is
defined as time-to-event of the composite endpoint of cardiovascular death,
ischemic stroke or myocardial infarction. The dataset includes those patients
enrolled in study between September 1996 and March 2006.
}
\keyword{datasets}
